Regenerate FreeFont manually
Encoding > Force Encoding > Unicode BMP

It builds with a crapton of warnings, but doesn't exhibit the issue
detailed in https://github.com/koreader/koreader/pull/4744#issuecomment-473113760

Note the doing a *Reencode* instead squishes the warnings, but
reintroduces the issues.

(We skip upstream's python shenanigans, which we don't need in these
cases anyway).